% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR.R
\name{BMR_female.Mifflin}
\alias{BMR_female.Mifflin}
\title{Calculate your Basal Metabolic Rate (BMR) using the Mifflin-St Jeor equation}
\usage{
BMR_female.Mifflin(x, y, z)
}
\arguments{
\item{weight}{in kg}

\item{height}{in cm}

\item{age}{age}
}
\value{
Your BMR calculation for female using Mifflin-St Jeor equation
}
\description{
BMR is the estimated amount of energy the body burns at rest to maintain normal bodily functions such as breathing, digestion, and heart rate.
BMR is often used as a starting point for determining a person’s total daily energy expenditure, which takes into account a person’s physical activity level in addition to their BMR.
}
\examples{
BMR_female.Mifflin(64,162.56,25)
}
