% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{generateMealPlan}
\alias{generateMealPlan}
\title{ChatGPT Generated Meal Plan}
\usage{
generateMealPlan(calories, protein, fats)
}
\arguments{
\item{fats}{}
}
\value{
Uses ChatGPT API to create a custom meal plan based on a users desired calories, protien, and fats
}
\description{
ChatGPT Generated Meal Plan
}
\examples{
generateMealPlan(1654, 90, 45)

}
