% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-conversion.R
\name{weight.kg}
\alias{weight.kg}
\title{Weight Conversion Pounds to Kilograms}
\usage{
weight.kg(x)
}
\arguments{
\item{x}{your weight in pounds}
}
\value{
inputted weight in kilograms
}
\description{
Weight Conversion Pounds to Kilograms
}
\examples{
weight.kg(125) should return 56.69
}
