% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-conversion.R
\name{weight.lb}
\alias{weight.lb}
\title{Weight Conversion Kilograms to Pounds}
\usage{
weight.lb(x)
}
\arguments{
\item{x}{weight in kilogram}
}
\value{
weight in pounds
}
\description{
Weight Conversion Kilograms to Pounds
}
\examples{
weight.lb(56.69) should return 125
}
